requirements.txt
anyio==3.6.2
asgiref==3.4.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
orjson==3.8.3
Pillow==8.3.2
pydantic==1.10.4
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spintax==1.0.4
sqlparse==0.4.2
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
